Add files via upload
This kernel is dedicated to finding the factors that lead a Haitian child toward depression and alter Haiti Now's future outreach to best support children in those areas that impact them the most. This will be found by predicting the depression level of the surveyed Haitian children and using the category coefficients to find the heaviest weighted variables. We will use demographic, household, and school information to gather insight.